Drafts backend (#131)
* Drafts API

+++ internal/apiserver/config/config.go
@@ -78,7 +78,7 @@ func New() *Configuration {
 
 	config.ValidationServerUrl = environment.Getenv("VALIDATION_SERVER_URL", config.ValidationServerUrl)
 
-	config.MaxInvoiceSize = int64(environment.ParseInt64("MAX_INVOICE_SIZE", config.MaxInvoiceSize))
+	config.MaxInvoiceSize = environment.ParseInt64("MAX_INVOICE_SIZE", config.MaxInvoiceSize)
 
 	log.Info("config.loaded")
